requirements-rag.txt
faiss-cpu
numpy
sentence-transformers
scikit-learn